requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.4.2
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
streamlit>=1.33.0
psycopg2-binary>=2.9.9
oracledb>=2.2.0
pandas>=2.2.1
numpy>=1.26.4
sentence-transformers>=2.5.1
apscheduler>=3.10.4
requests>=2.31.0
